backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cors==1.0.1
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
fastapi==0.121.0
filelock==3.20.0
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
multidict==6.7.0
numpy==2.3.4
openai==2.7.1
packaging==25.0
postgrest==2.23.3
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.23.3
requests==2.32.5
requests-file==3.0.1
sniffio==1.3.1
starlette==0.49.3
storage3==2.23.3
StrEnum==0.4.15
supabase==2.23.3
supabase-auth==2.23.3
supabase-functions==2.23.3
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zope.event==6.0
zope.interface==8.0.1
